Update exec to only exec if package is actually installed.

+++ manifests/ubuntu.pp
@@ -39,7 +39,8 @@ class datadog_agent::ubuntu(
   # $facts hash - for compatibility with puppet3.x default behavior
   if $::apt_agent6_beta_repo and $agent_version == 'latest' {
     exec { 'datadog_apt-get_remove_agent6':
-      command     => '/usr/bin/apt-get remove -y -q datadog-agent',
+      command => '/usr/bin/apt-get remove -y -q datadog-agent',
+      onlyif  => '/usr/bin/dpkg -l datadog-agent > /dev/null'
     }
   } else {
     exec { 'datadog_apt-get_remove_agent6':
